+++ src/main/java/org/cicirello/ds/BinaryHeap.java
@@ -83,7 +83,7 @@ public final class BinaryHeap<E>
   private PriorityQueueNode.Integer<E>[] buffer;
   private int size;
   private final HashMap<E, java.lang.Integer> index;
-  private final PriorityComparator compare;
+  private final Prioritizer compare;
   private final int extreme;
 
   /** The default initial capacity. */
@@ -96,7 +96,7 @@ public final class BinaryHeap<E>
    * @param initialCapacity The initial capacity, which must be positive.
    */
   private BinaryHeap(int initialCapacity) {
-    this(initialCapacity, (p1, p2) -> p1 < p2);
+    this(initialCapacity, new MinOrder());
   }
 
   /* PRIVATE: Use factory methods for creation.
@@ -105,12 +105,11 @@ public final class BinaryHeap<E>
    *
    * @param initialCapacity The initial capacity, which must be positive.
    */
-  private BinaryHeap(int initialCapacity, PriorityComparator compare) {
+  private BinaryHeap(int initialCapacity, Prioritizer compare) {
     this.compare = compare;
     buffer = allocate(initialCapacity);
     index = new HashMap<E, java.lang.Integer>();
-    extreme =
-        compare.belongsAbove(0, 1) ? java.lang.Integer.MAX_VALUE : java.lang.Integer.MIN_VALUE;
+    extreme = compare.comesBefore(0, 1) ? java.lang.Integer.MAX_VALUE : java.lang.Integer.MIN_VALUE;
   }
 
   /* PRIVATE: Use factory methods for creation.
@@ -124,7 +123,7 @@ public final class BinaryHeap<E>
    * one pair in initialElements contains the same element.
    */
   private BinaryHeap(Collection<PriorityQueueNode.Integer<E>> initialElements) {
-    this(initialElements, (p1, p2) -> p1 < p2);
+    this(initialElements, new MinOrder());
   }
 
   /* PRIVATE: Use factory methods for creation.
@@ -138,7 +137,7 @@ public final class BinaryHeap<E>
    * one pair in initialElements contains the same element.
    */
   private BinaryHeap(
-      Collection<PriorityQueueNode.Integer<E>> initialElements, PriorityComparator compare) {
+      Collection<PriorityQueueNode.Integer<E>> initialElements, Prioritizer compare) {
     this(initialElements.size(), compare);
     for (PriorityQueueNode.Integer<E> element : initialElements) {
       if (index.containsKey(element.element)) {
@@ -221,7 +220,7 @@ public final class BinaryHeap<E>
    * @return an empty BinaryHeap with a maximum-priority-first-out priority order
    */
   public static <E> BinaryHeap<E> createMaxHeap() {
-    return new BinaryHeap<E>(DEFAULT_INITIAL_CAPACITY, (p1, p2) -> p1 > p2);
+    return new BinaryHeap<E>(DEFAULT_INITIAL_CAPACITY, new MaxOrder());
   }
 
   /**
@@ -236,7 +235,7 @@ public final class BinaryHeap<E>
   public static <E> BinaryHeap<E> createMaxHeap(int initialCapacity) {
     if (initialCapacity <= 0)
       throw new IllegalArgumentException("Initial capacity must be positive.");
-    return new BinaryHeap<E>(initialCapacity, (p1, p2) -> p1 > p2);
+    return new BinaryHeap<E>(initialCapacity, new MaxOrder());
   }
 
   /**
@@ -255,7 +254,7 @@ public final class BinaryHeap<E>
     if (initialElements.size() < 1) {
       throw new IllegalArgumentException("initialElements is empty");
     }
-    return new BinaryHeap<E>(initialElements, (p1, p2) -> p1 > p2);
+    return new BinaryHeap<E>(initialElements, new MaxOrder());
   }
 
   /**
@@ -295,11 +294,11 @@ public final class BinaryHeap<E>
   public final boolean change(E element, int priority) {
     if (!offer(element, priority)) {
       int i = index.get(element);
-      if (compare.belongsAbove(priority, buffer[i].value)) {
+      if (compare.comesBefore(priority, buffer[i].value)) {
         buffer[i].value = priority;
         percolateUp(i);
         return true;
-      } else if (compare.belongsAbove(buffer[i].value, priority)) {
+      } else if (compare.comesBefore(buffer[i].value, priority)) {
         buffer[i].value = priority;
         percolateDown(i);
         return true;
@@ -332,7 +331,7 @@ public final class BinaryHeap<E>
     Integer where = index.get(element);
     if (where != null) {
       int i = where;
-      if (compare.belongsAbove(buffer[i].value, priority)) {
+      if (compare.comesBefore(buffer[i].value, priority)) {
         buffer[i].value = priority;
         percolateDown(i);
         return true;
@@ -368,7 +367,7 @@ public final class BinaryHeap<E>
       @SuppressWarnings("unchecked")
       BinaryHeap<E> casted = (BinaryHeap<E>) other;
       if (size != casted.size) return false;
-      if (compare.belongsAbove(0, 1) != casted.compare.belongsAbove(0, 1)) return false;
+      if (compare.comesBefore(0, 1) != casted.compare.comesBefore(0, 1)) return false;
       for (int i = 0; i < size; i++) {
         if (!buffer[i].element.equals(casted.buffer[i].element)) return false;
         if (casted.buffer[i].value != buffer[i].value) return false;
@@ -412,7 +411,7 @@ public final class BinaryHeap<E>
    */
   @Override
   public boolean merge(BinaryHeap<E> other) {
-    if (compare.belongsAbove(0, 1) != other.compare.belongsAbove(0, 1)) {
+    if (compare.comesBefore(0, 1) != other.compare.comesBefore(0, 1)) {
       throw new IllegalArgumentException("this and other follow different priority-order");
     }
     if (size + other.size() > buffer.length) {
@@ -552,7 +551,7 @@ public final class BinaryHeap<E>
     Integer where = index.get(element);
     if (where != null) {
       int i = where;
-      if (compare.belongsAbove(priority, buffer[i].value)) {
+      if (compare.comesBefore(priority, buffer[i].value)) {
         buffer[i].value = priority;
         percolateUp(i);
         return true;
@@ -581,9 +580,9 @@ public final class BinaryHeap<E>
       buffer[size] = null;
       index.put(buffer[i].element, i);
       // percolate in relevant direction
-      if (compare.belongsAbove(buffer[i].value, removedElementPriority)) {
+      if (compare.comesBefore(buffer[i].value, removedElementPriority)) {
         percolateUp(i);
-      } else if (compare.belongsAbove(removedElementPriority, buffer[i].value)) {
+      } else if (compare.comesBefore(removedElementPriority, buffer[i].value)) {
         percolateDown(i);
       }
     } else {
@@ -734,11 +733,11 @@ public final class BinaryHeap<E>
     int left;
     while ((left = (i << 1) + 1) < size) {
       int smallest = i;
-      if (compare.belongsAbove(buffer[left].value, buffer[i].value)) {
+      if (compare.comesBefore(buffer[left].value, buffer[i].value)) {
         smallest = left;
       }
       int right = left + 1;
-      if (right < size && compare.belongsAbove(buffer[right].value, buffer[smallest].value)) {
+      if (right < size && compare.comesBefore(buffer[right].value, buffer[smallest].value)) {
         smallest = right;
       }
       if (smallest != i) {
@@ -756,7 +755,7 @@ public final class BinaryHeap<E>
 
   private void percolateUp(int i) {
     int parent;
-    while (i > 0 && compare.belongsAbove(buffer[i].value, buffer[parent = (i - 1) >> 1].value)) {
+    while (i > 0 && compare.comesBefore(buffer[i].value, buffer[parent = (i - 1) >> 1].value)) {
       PriorityQueueNode.Integer<E> temp = buffer[i];
       buffer[i] = buffer[parent];
       buffer[parent] = temp;
@@ -799,12 +798,6 @@ public final class BinaryHeap<E>
     for (int i = (size >> 1) - 1; i >= 0; i--) {
       percolateDown(i);
     }
-  }
-
-  @FunctionalInterface
-  private static interface PriorityComparator {
-
-    boolean belongsAbove(int p1, int p2);
   }
 
   private class BinaryHeapIterator implements Iterator<PriorityQueueNode.Integer<E>> {