+++ src/main/java/org/cicirello/ds/BinaryHeapDouble.java
@@ -83,7 +83,7 @@ public final class BinaryHeapDouble<E>
   private PriorityQueueNode.Double<E>[] buffer;
   private int size;
   private final HashMap<E, java.lang.Integer> index;
-  private final PriorityComparator compare;
+  private final Prioritizer compare;
   private final double extreme;
 
   /** The default initial capacity. */
@@ -96,7 +96,7 @@ public final class BinaryHeapDouble<E>
    * @param initialCapacity The initial capacity, which must be positive.
    */
   private BinaryHeapDouble(int initialCapacity) {
-    this(initialCapacity, (p1, p2) -> p1 < p2);
+    this(initialCapacity, new MinOrder());
   }
 
   /* PRIVATE: Use factory methods for creation.
@@ -105,12 +105,12 @@ public final class BinaryHeapDouble<E>
    *
    * @param initialCapacity The initial capacity, which must be positive.
    */
-  private BinaryHeapDouble(int initialCapacity, PriorityComparator compare) {
+  private BinaryHeapDouble(int initialCapacity, Prioritizer compare) {
     this.compare = compare;
     buffer = allocate(initialCapacity);
     index = new HashMap<E, java.lang.Integer>();
     extreme =
-        compare.belongsAbove(0, 1)
+        compare.comesBefore(0, 1)
             ? java.lang.Double.POSITIVE_INFINITY
             : java.lang.Double.NEGATIVE_INFINITY;
   }
@@ -126,7 +126,7 @@ public final class BinaryHeapDouble<E>
    * one pair in initialElements contains the same element.
    */
   private BinaryHeapDouble(Collection<PriorityQueueNode.Double<E>> initialElements) {
-    this(initialElements, (p1, p2) -> p1 < p2);
+    this(initialElements, new MinOrder());
   }
 
   /* PRIVATE: Use factory methods for creation.
@@ -140,7 +140,7 @@ public final class BinaryHeapDouble<E>
    * one pair in initialElements contains the same element.
    */
   private BinaryHeapDouble(
-      Collection<PriorityQueueNode.Double<E>> initialElements, PriorityComparator compare) {
+      Collection<PriorityQueueNode.Double<E>> initialElements, Prioritizer compare) {
     this(initialElements.size(), compare);
     for (PriorityQueueNode.Double<E> element : initialElements) {
       if (index.containsKey(element.element)) {
@@ -223,7 +223,7 @@ public final class BinaryHeapDouble<E>
    * @return an empty BinaryHeapDouble with a maximum-priority-first-out priority order
    */
   public static <E> BinaryHeapDouble<E> createMaxHeap() {
-    return new BinaryHeapDouble<E>(DEFAULT_INITIAL_CAPACITY, (p1, p2) -> p1 > p2);
+    return new BinaryHeapDouble<E>(DEFAULT_INITIAL_CAPACITY, new MaxOrder());
   }
 
   /**
@@ -238,7 +238,7 @@ public final class BinaryHeapDouble<E>
   public static <E> BinaryHeapDouble<E> createMaxHeap(int initialCapacity) {
     if (initialCapacity <= 0)
       throw new IllegalArgumentException("Initial capacity must be positive.");
-    return new BinaryHeapDouble<E>(initialCapacity, (p1, p2) -> p1 > p2);
+    return new BinaryHeapDouble<E>(initialCapacity, new MaxOrder());
   }
 
   /**
@@ -257,7 +257,7 @@ public final class BinaryHeapDouble<E>
     if (initialElements.size() < 1) {
       throw new IllegalArgumentException("initialElements is empty");
     }
-    return new BinaryHeapDouble<E>(initialElements, (p1, p2) -> p1 > p2);
+    return new BinaryHeapDouble<E>(initialElements, new MaxOrder());
   }
 
   /**
@@ -297,11 +297,11 @@ public final class BinaryHeapDouble<E>
   public final boolean change(E element, double priority) {
     if (!offer(element, priority)) {
       int i = index.get(element);
-      if (compare.belongsAbove(priority, buffer[i].value)) {
+      if (compare.comesBefore(priority, buffer[i].value)) {
         buffer[i].value = priority;
         percolateUp(i);
         return true;
-      } else if (compare.belongsAbove(buffer[i].value, priority)) {
+      } else if (compare.comesBefore(buffer[i].value, priority)) {
         buffer[i].value = priority;
         percolateDown(i);
         return true;
@@ -334,7 +334,7 @@ public final class BinaryHeapDouble<E>
     Integer where = index.get(element);
     if (where != null) {
       int i = where;
-      if (compare.belongsAbove(buffer[i].value, priority)) {
+      if (compare.comesBefore(buffer[i].value, priority)) {
         buffer[i].value = priority;
         percolateDown(i);
         return true;
@@ -370,7 +370,7 @@ public final class BinaryHeapDouble<E>
       @SuppressWarnings("unchecked")
       BinaryHeapDouble<E> casted = (BinaryHeapDouble<E>) other;
       if (size != casted.size) return false;
-      if (compare.belongsAbove(0, 1) != casted.compare.belongsAbove(0, 1)) return false;
+      if (compare.comesBefore(0, 1) != casted.compare.comesBefore(0, 1)) return false;
       for (int i = 0; i < size; i++) {
         if (!buffer[i].element.equals(casted.buffer[i].element)) return false;
         if (casted.buffer[i].value != buffer[i].value) return false;
@@ -414,7 +414,7 @@ public final class BinaryHeapDouble<E>
    */
   @Override
   public boolean merge(BinaryHeapDouble<E> other) {
-    if (compare.belongsAbove(0, 1) != other.compare.belongsAbove(0, 1)) {
+    if (compare.comesBefore(0, 1) != other.compare.comesBefore(0, 1)) {
       throw new IllegalArgumentException("this and other follow different priority-order");
     }
     if (size + other.size() > buffer.length) {
@@ -554,7 +554,7 @@ public final class BinaryHeapDouble<E>
     Integer where = index.get(element);
     if (where != null) {
       int i = where;
-      if (compare.belongsAbove(priority, buffer[i].value)) {
+      if (compare.comesBefore(priority, buffer[i].value)) {
         buffer[i].value = priority;
         percolateUp(i);
         return true;
@@ -583,9 +583,9 @@ public final class BinaryHeapDouble<E>
       buffer[size] = null;
       index.put(buffer[i].element, i);
       // percolate in relevant direction
-      if (compare.belongsAbove(buffer[i].value, removedElementPriority)) {
+      if (compare.comesBefore(buffer[i].value, removedElementPriority)) {
         percolateUp(i);
-      } else if (compare.belongsAbove(removedElementPriority, buffer[i].value)) {
+      } else if (compare.comesBefore(removedElementPriority, buffer[i].value)) {
         percolateDown(i);
       }
     } else {
@@ -736,11 +736,11 @@ public final class BinaryHeapDouble<E>
     int left;
     while ((left = (i << 1) + 1) < size) {
       int smallest = i;
-      if (compare.belongsAbove(buffer[left].value, buffer[i].value)) {
+      if (compare.comesBefore(buffer[left].value, buffer[i].value)) {
         smallest = left;
       }
       int right = left + 1;
-      if (right < size && compare.belongsAbove(buffer[right].value, buffer[smallest].value)) {
+      if (right < size && compare.comesBefore(buffer[right].value, buffer[smallest].value)) {
         smallest = right;
       }
       if (smallest != i) {
@@ -758,7 +758,7 @@ public final class BinaryHeapDouble<E>
 
   private void percolateUp(int i) {
     int parent;
-    while (i > 0 && compare.belongsAbove(buffer[i].value, buffer[parent = (i - 1) >> 1].value)) {
+    while (i > 0 && compare.comesBefore(buffer[i].value, buffer[parent = (i - 1) >> 1].value)) {
       PriorityQueueNode.Double<E> temp = buffer[i];
       buffer[i] = buffer[parent];
       buffer[parent] = temp;
@@ -801,12 +801,6 @@ public final class BinaryHeapDouble<E>
     for (int i = (size >> 1) - 1; i >= 0; i--) {
       percolateDown(i);
     }
-  }
-
-  @FunctionalInterface
-  private static interface PriorityComparator {
-
-    boolean belongsAbove(double p1, double p2);
   }
 
   private class BinaryHeapDoubleIterator implements Iterator<PriorityQueueNode.Double<E>> {