+++ src/main/java/org/cicirello/ds/SimpleBinaryHeapDouble.java
@@ -83,7 +83,7 @@ public final class SimpleBinaryHeapDouble<E>
 
   private PriorityQueueNode.Double<E>[] buffer;
   private int size;
-  private final PriorityComparator compare;
+  private final Prioritizer compare;
   private final double extreme;
 
   /** The default initial capacity. */
@@ -96,7 +96,7 @@ public final class SimpleBinaryHeapDouble<E>
    * @param initialCapacity The initial capacity, which must be positive.
    */
   private SimpleBinaryHeapDouble(int initialCapacity) {
-    this(initialCapacity, (p1, p2) -> p1 < p2);
+    this(initialCapacity, new MinOrder());
   }
 
   /* PRIVATE: Use factory methods for creation.
@@ -105,11 +105,11 @@ public final class SimpleBinaryHeapDouble<E>
    *
    * @param initialCapacity The initial capacity, which must be positive.
    */
-  private SimpleBinaryHeapDouble(int initialCapacity, PriorityComparator compare) {
+  private SimpleBinaryHeapDouble(int initialCapacity, Prioritizer compare) {
     this.compare = compare;
     buffer = allocate(initialCapacity);
     extreme =
-        compare.belongsAbove(0, 1)
+        compare.comesBefore(0, 1)
             ? java.lang.Double.POSITIVE_INFINITY
             : java.lang.Double.NEGATIVE_INFINITY;
   }
@@ -124,7 +124,7 @@ public final class SimpleBinaryHeapDouble<E>
    * @throws IllegalArgumentException if initialElements is empty.
    */
   private SimpleBinaryHeapDouble(Collection<PriorityQueueNode.Double<E>> initialElements) {
-    this(initialElements, (p1, p2) -> p1 < p2);
+    this(initialElements, new MinOrder());
   }
 
   /* PRIVATE: Use factory methods for creation.
@@ -137,7 +137,7 @@ public final class SimpleBinaryHeapDouble<E>
    * @throws IllegalArgumentException if initialElements is empty.
    */
   private SimpleBinaryHeapDouble(
-      Collection<PriorityQueueNode.Double<E>> initialElements, PriorityComparator compare) {
+      Collection<PriorityQueueNode.Double<E>> initialElements, Prioritizer compare) {
     this(initialElements.size(), compare);
     for (PriorityQueueNode.Double<E> element : initialElements) {
       buffer[size] = element.copy();
@@ -214,7 +214,7 @@ public final class SimpleBinaryHeapDouble<E>
    * @return an empty SimpleBinaryHeapDouble with a maximum-priority-first-out priority order
    */
   public static <E> SimpleBinaryHeapDouble<E> createMaxHeap() {
-    return new SimpleBinaryHeapDouble<E>(DEFAULT_INITIAL_CAPACITY, (p1, p2) -> p1 > p2);
+    return new SimpleBinaryHeapDouble<E>(DEFAULT_INITIAL_CAPACITY, new MaxOrder());
   }
 
   /**
@@ -229,7 +229,7 @@ public final class SimpleBinaryHeapDouble<E>
   public static <E> SimpleBinaryHeapDouble<E> createMaxHeap(int initialCapacity) {
     if (initialCapacity <= 0)
       throw new IllegalArgumentException("Initial capacity must be positive.");
-    return new SimpleBinaryHeapDouble<E>(initialCapacity, (p1, p2) -> p1 > p2);
+    return new SimpleBinaryHeapDouble<E>(initialCapacity, new MaxOrder());
   }
 
   /**
@@ -247,7 +247,7 @@ public final class SimpleBinaryHeapDouble<E>
     if (initialElements.size() < 1) {
       throw new IllegalArgumentException("initialElements is empty");
     }
-    return new SimpleBinaryHeapDouble<E>(initialElements, (p1, p2) -> p1 > p2);
+    return new SimpleBinaryHeapDouble<E>(initialElements, new MaxOrder());
   }
 
   @Override
@@ -290,11 +290,11 @@ public final class SimpleBinaryHeapDouble<E>
   public final boolean change(E element, double priority) {
     int i = find(element);
     if (i >= 0) {
-      if (compare.belongsAbove(priority, buffer[i].value)) {
+      if (compare.comesBefore(priority, buffer[i].value)) {
         buffer[i].value = priority;
         percolateUp(i);
         return true;
-      } else if (compare.belongsAbove(buffer[i].value, priority)) {
+      } else if (compare.comesBefore(buffer[i].value, priority)) {
         buffer[i].value = priority;
         percolateDown(i);
         return true;
@@ -351,7 +351,7 @@ public final class SimpleBinaryHeapDouble<E>
   public final boolean demote(E element, double priority) {
     int i = find(element);
     if (i >= 0) {
-      if (compare.belongsAbove(buffer[i].value, priority)) {
+      if (compare.comesBefore(buffer[i].value, priority)) {
         buffer[i].value = priority;
         percolateDown(i);
         return true;
@@ -387,7 +387,7 @@ public final class SimpleBinaryHeapDouble<E>
       @SuppressWarnings("unchecked")
       SimpleBinaryHeapDouble<E> casted = (SimpleBinaryHeapDouble<E>) other;
       if (size != casted.size) return false;
-      if (compare.belongsAbove(0, 1) != casted.compare.belongsAbove(0, 1)) return false;
+      if (compare.comesBefore(0, 1) != casted.compare.comesBefore(0, 1)) return false;
       for (int i = 0; i < size; i++) {
         if (!buffer[i].element.equals(casted.buffer[i].element)) return false;
         if (casted.buffer[i].value != buffer[i].value) return false;
@@ -431,7 +431,7 @@ public final class SimpleBinaryHeapDouble<E>
    */
   @Override
   public boolean merge(SimpleBinaryHeapDouble<E> other) {
-    if (compare.belongsAbove(0, 1) != other.compare.belongsAbove(0, 1)) {
+    if (compare.comesBefore(0, 1) != other.compare.comesBefore(0, 1)) {
       throw new IllegalArgumentException("this and other follow different priority-order");
     }
     if (size + other.size() > buffer.length) {
@@ -533,7 +533,7 @@ public final class SimpleBinaryHeapDouble<E>
   public final boolean promote(E element, double priority) {
     int i = find(element);
     if (i >= 0) {
-      if (compare.belongsAbove(priority, buffer[i].value)) {
+      if (compare.comesBefore(priority, buffer[i].value)) {
         buffer[i].value = priority;
         percolateUp(i);
         return true;
@@ -560,9 +560,9 @@ public final class SimpleBinaryHeapDouble<E>
       buffer[i] = buffer[size];
       buffer[size] = null;
       // percolate in relevant direction
-      if (compare.belongsAbove(buffer[i].value, removedElementPriority)) {
+      if (compare.comesBefore(buffer[i].value, removedElementPriority)) {
         percolateUp(i);
-      } else if (compare.belongsAbove(removedElementPriority, buffer[i].value)) {
+      } else if (compare.comesBefore(removedElementPriority, buffer[i].value)) {
         percolateDown(i);
       }
     } else {
@@ -714,11 +714,11 @@ public final class SimpleBinaryHeapDouble<E>
     int left;
     while ((left = (i << 1) + 1) < size) {
       int smallest = i;
-      if (compare.belongsAbove(buffer[left].value, buffer[i].value)) {
+      if (compare.comesBefore(buffer[left].value, buffer[i].value)) {
         smallest = left;
       }
       int right = left + 1;
-      if (right < size && compare.belongsAbove(buffer[right].value, buffer[smallest].value)) {
+      if (right < size && compare.comesBefore(buffer[right].value, buffer[smallest].value)) {
         smallest = right;
       }
       if (smallest != i) {
@@ -734,7 +734,7 @@ public final class SimpleBinaryHeapDouble<E>
 
   private void percolateUp(int i) {
     int parent;
-    while (i > 0 && compare.belongsAbove(buffer[i].value, buffer[parent = (i - 1) >> 1].value)) {
+    while (i > 0 && compare.comesBefore(buffer[i].value, buffer[parent = (i - 1) >> 1].value)) {
       PriorityQueueNode.Double<E> temp = buffer[i];
       buffer[i] = buffer[parent];
       buffer[parent] = temp;
@@ -774,11 +774,6 @@ public final class SimpleBinaryHeapDouble<E>
     for (int i = (size >> 1) - 1; i >= 0; i--) {
       percolateDown(i);
     }
-  }
-
-  @FunctionalInterface
-  private static interface PriorityComparator {
-    boolean belongsAbove(double p1, double p2);
   }
 
   private class SimpleBinaryHeapDoubleIterator implements Iterator<PriorityQueueNode.Double<E>> {